update(convtasnet): refactor a bunch of stuff + script to generate compiled models and run benchmark inferences

+++ include/ModelInferenceMethods/GeneralInferenceParams.hpp
@@ -11,5 +11,5 @@ struct GeneralInferenceParams {
     bool                      profiling;
     std::string               profiling_file;
     size_t                    warm_up_buffers;
-    int64_t                   dsp_sample_rate;
+    int                       dsp_sample_rate;
 };
